Increasing possible media count per folder fixes #559

+++ src/backend/model/database/sql/enitites/DirectoryEntity.ts
@@ -61,7 +61,7 @@ export class DirectoryEntity
 
   isPartial?: boolean;
 
-  @Column('smallint', { unsigned: true })
+  @Column('mediumint', { unsigned: true })
   mediaCount: number;
 
   @Index()
